Add local data mirror and auto dataset sizing
Add web-accessible CIFAR-10 mirror path and update fetch script
to prefer local mirror. Enable automatic train/eval sizing from
dataset metadata across multiple trainers.

- cifar/ki-local.h: add www/data paths for web deployment
- fetch_cifar10.sh: prefer local mirror, derive BASE from script dir
- mnist/ki-local.h: cast realloc/memcpy sizes to size_t
- mnist/mlp-bin32-hebbian-trn.c: default trainN/evalN to 0, auto-detect from dataset
- mnist/mlp-bin32-otto-trn.c: use n_train+n_eval as upper bound instead of num_images
- mnist/mlp-flt32-adam-trn.c: default to 0 and auto-detect, use n_train+n_eval bound

Prevents overflow on 32-bit and avoids hard-coded sizes when dataset
metadata is available. Improves robustness for dry-run and web setups.

+++ mnist/ki-local.h
@@ -233,8 +233,8 @@ static int ki_mnist_read(ki_MNISTData *out) {
                     for (size_t i = 0; i < te_npix; i++)
                         out->X[old_npix + i] = ((float)te_raw[16 + i] / 255.0f) * 2.0f - 1.0f;
                     /* Realloc + copy labels */
-                    out->y = (uint8_t *)realloc(out->y, num_img + te_num);
-                    memcpy(out->y + num_img, tel + 8, te_num);
+                    out->y = (uint8_t *)realloc(out->y, (size_t)num_img + (size_t)te_num);
+                    memcpy(out->y + num_img, tel + 8, (size_t)te_num);
                     out->num_images += te_num;
                     out->n_eval = te_num;
                 }
